ui/pyproject.toml
[project]
name = "ui-sandbox"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "pyarrow>=22.0.0",
    "streamlit>=1.50.0",
]
